Introduce dense leaf nodes.
If a set of voxels covered by a tree node precisely matches the bounding box of
that node, this node is considered a "dense leaf" now. A "dense leaf" is a
special kind of leaf. It can have any number of dots in it, not just up to
maximum number of VOX_MAX_DOTS, as long as these dots satisfy the criteria I
mentioned. Dense leafs also require a minimal amount of memory, just enough to
store its bounding box and flags.
This improvement saves a lot of memory and also optimizes a search algorithm a
tiny bit. The latter is because of dense nodes are usually "covered" by
"ordinary" ones. But you can benefit from them when rendering a solid cube, for
example.