Dirty fix for ray/tree intersection special case.
There is a special case in vox_ray_tree_intersection() function. This is when
the ray's origin lays on any dividing plane of any tree node. In this case the
starting subspace index must be calculated with respect to the ray's direction
vector, not just by calling get_subspace_idx(). This is a dirty fix which fixes
the problem for rays which are not parallel for that dividing plane.