Do not block rendering when rebuilding the tree.
Since there is a very small amount of insertions/deletions compared to searches,
I've decided to completely stop the renderer when the tree is changing. On the
other hand tree rebuilding does not destructively modify the old tree, so it can
be run in parallel to renderer.

This is archieved with GCD queues and groups.