Fix probable memory allocation error.
From FreeBSD's man page on aligned_alloc:
"Behaviour is undefined if size is not an integral multiple of alignment". As we
do not make any guarantees on size of objects containing fields of type vox_dot
or vox_quad when building with SSE_INTRIN=OFF, use malloc() in that case.