Working point light concept.
User can add a point light to the scene with vox_insert_point_light() and delete
with vox_delete_point_light(). Currently lights are black-and-white only (color
argument is ignored) and does not follow any realistic math model. It's just a
concept. Demo application inserts a light in the camera's origin.